LayoutTests/imported/w3c/web-platform-tests/svg/text/scripted/getcharnumatposition-slr.tentative-expected.txt

abc
abc

FAIL sideways-lr assert_equals: expected -1 but got 0
